requirements/production.txt
Django
djangorestframework
requests
beautifulsoup4
pyperclip
python-social-auth
django-pipeline
celery[redis]
redis
flower
dj-database-url
psycopg2
honcho
gunicorn
django-storages-redux
boto
